Manual boxing edited

+++ Wrap.java
@@ -1,7 +1,7 @@
 //Boxing and unboxing with a type wrapper
 class Wrap {
 	public static void main(String args[]) {
-		Integer iOb = new Integer(100); //boxing
+		Integer iOb = new Integer(247); //boxing
 		
 		int i = iOb.intValue(); //unboxing
